requirements.txt
pyramid-jwtauth==0.1.0
pyramid==1.6a1
pyramid-sockjs==0.3.9
pyparsing==2.0.3
pyramid-dogpile-cache==0.0.4
